ml/requirements.txt
ultralytics>=8.4.11
torch>=2.0.0
torchvision>=0.15.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
